+++ src/main.rs
@@ -20,6 +20,7 @@ use mitra::http::{
     create_default_headers_middleware,
     json_error_handler,
     multiquery_config,
+    ratelimit_config,
 };
 use mitra::init::initialize_app;
 use mitra::job_queue::scheduler;
@@ -101,6 +102,9 @@ async fn main() -> std::io::Result<()> {
     // Mutex is used to make server process incoming activities sequentially
     let inbox_mutex = web::Data::new(Mutex::new(()));
 
+    // One request per 5 seconds
+    let acct_resolve_ratelimit_config = ratelimit_config(2, 30);
+
     let http_server = HttpServer::new(move || {
         let cors_config = match config.environment {
             Environment::Development => {
@@ -162,7 +166,7 @@ async fn main() -> std::io::Result<()> {
                 config.media_dir(),
             ))
             .service(oauth_api_scope())
-            .service(account_api_scope())
+            .service(account_api_scope(&acct_resolve_ratelimit_config))
             .service(application_api_scope())
             .service(custom_emoji_api_scope())
             .service(directory_api_scope())