+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Add rate-limited variant of `/api/v1/accounts/search` endpoint.
+
 ### Changed
 
 - Allow `gemini:` and `monero:` links in post content.